+++ app/src/tab/media_series.cpp
@@ -137,16 +137,16 @@ MediaSeries::MediaSeries(const jellyfin::Item& item) : seriesId(item.Id) {
     this->nextUp->registerCell("Cell", VideoCardCell::create);
 
     this->registerAction(
-        "上一项", brls::ControllerButton::BUTTON_LB,
-        [this](brls::View* view) -> bool {
+        "main/play/next"_i18n, brls::BUTTON_LB,
+        [this](brls::View* view) {
             tabFrame->focus2LastTab();
             return true;
         },
         true);
 
     this->registerAction(
-        "下一项", brls::ControllerButton::BUTTON_RB,
-        [this](brls::View* view) -> bool {
+        "main/play/pref"_i18n, brls::BUTTON_RB,
+        [this](brls::View* view) {
             tabFrame->focus2NextTab();
             return true;
         },